requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
databases==0.7.0
pydantic-settings==2.0.2
pandas==2.0.3
pytest===7.4.0
alembic==1.11.1
motor==3.2.0
pydantic[email]
kafka-python==2.0.2
confluent-kafka==2.2.0
python-dotenv==1.0.0
pymongo==4.4.1